Migrate import and export endpoints to new modules
Creates the new import and export service and controller modules and
migrates the upload and download zip logic to these modules. Does not
deprecate the old endpoints
