+++ oxen-rust/src/lib/src/api/client/dir.rs
@@ -104,7 +104,7 @@ pub async fn download_dir_as_zip(
     let revision = revision.as_ref().to_string();
     let directory = directory.as_ref().to_string_lossy().to_string();
     let local_path = local_path.as_ref();
-    let uri = format!("/dir/download_zip/{revision}/{directory}");
+    let uri = format!("/export/download_zip/{revision}/{directory}");
     let url = api::endpoint::url_from_repo(remote_repo, &uri)?;
 
     let client = client::new_for_url(&url)?;
