+++ oxen-rust/src/server/src/main.rs
@@ -170,8 +170,11 @@ const SUPPORT: &str = "
         // Files (Repository)
         crate::controllers::file::get,
         crate::controllers::file::put,
-        crate::controllers::file::upload_zip,
-        crate::controllers::file::import,
+        // Import
+        crate::controllers::import::upload_zip,
+        crate::controllers::import::import,
+        // Export
+        crate::controllers::export::download_zip,
         // DataFrames
         crate::controllers::data_frames::get,
         crate::controllers::data_frames::index,
@@ -218,8 +221,8 @@ const SUPPORT: &str = "
             // Upload & Request Bodies
             crate::controllers::workspaces::files::FileUpload,
             crate::controllers::file::FileUploadBody,
-            crate::controllers::file::ZipUploadBody,
-            crate::controllers::file::ImportFileBody,
+            crate::controllers::import::ZipUploadBody,
+            crate::controllers::import::ImportFileBody,
             FromDirectoryRequest,
             // Metadata Schemas
             EMetadataEntryResponseView,
