+++ oxen-rust/src/server/src/routes.rs
@@ -31,8 +31,10 @@ pub fn config(cfg: &mut web::ServiceConfig) {
                 .service(services::compare())
                 .service(services::data_frames())
                 .service(services::dir())
+                .service(services::export())
                 .service(services::file())
                 .service(services::fork())
+                .service(services::import())
                 .service(services::merge())
                 .service(services::meta())
                 .service(services::prune())
