+++ oxen-rust/src/server/src/services.rs
@@ -6,8 +6,10 @@ pub mod commits_db;
 pub mod compare;
 pub mod data_frames;
 pub mod dir;
+pub mod export;
 pub mod file;
 pub mod fork;
+pub mod import;
 pub mod merge;
 pub mod meta;
 pub mod prune;
@@ -29,8 +31,10 @@ pub use commits_db::commits_db;
 pub use compare::compare;
 pub use data_frames::data_frames;
 pub use dir::dir;
+pub use export::export;
 pub use file::file;
 pub use fork::fork;
+pub use import::import;
 pub use merge::merge;
 pub use meta::meta;
 pub use prune::prune;
