+++ oxen-rust/src/server/src/services/dir.rs
@@ -4,13 +4,5 @@ use actix_web::Scope;
 use crate::controllers;
 
 pub fn dir() -> Scope {
-    web::scope("/dir")
-        .route(
-            "/download/{resource:.*}",
-            web::get().to(controllers::dir::get),
-        )
-        .route(
-            "/download_zip/{resource:.*}",
-            web::get().to(controllers::dir::download_dir_as_zip),
-        )
+    web::scope("/dir").route("/{resource:.*}", web::get().to(controllers::dir::get))
 }
